Address PR review comments
- agent.agency: re-add the `map` import that the staged change
  accidentally removed (the call site still uses it).
- compile.ts: drop the `./${importPath}` rewrite in checkImportPolicy.
  importKind() already classifies *.agency paths as "local"; the rewrite
  poisoned glob matching (e.g. "./helpers.agency" became
  "././helpers.agency").
- utils.ts: make safeDeleteFile/safeDeleteDirectory best-effort. They
  now use rmSync({force:true}) and catch errors, so a cleanup hiccup in
  a finally block can't mask the real error.
- compile.ts: remove the deprecated `restrictImports` field entirely.
  Update all in-tree callers to `imports: { allowKinds: ["stdlib"] }`.
- typecheck.ts (new): extract TypeCheckDiagnostic / TypeCheckReport /
  typeCheckSource into their own module. compile.ts re-exports them
  for backward compat with stdlib/agency.ts.
- typecheck.ts: simplify the per-severity partition into a single pass
  instead of two .filter() calls. severity is always "error" | "warning"
  (verified against lib/typeChecker/types.ts).
- agency.ts: rewrite the nested ternary in resolveInSandbox as a plain
  if/else for readability.
- agency.agency: move _getImports / _getFunctions / _getGraphNodes out
  of TS into Agency one-liners that call getNodesOfType directly.
- builtins.ts: derive WriteMode from VALID_WRITE_MODES (single source of
  truth).
- utils.ts: add `uniqBy(arr, keyFn)`. synthesizer.ts uses it to replace
  the inline JSON.stringify-keyed Map dedup.
- importPaths.ts: drop the lazy-load picomatch wrapper; just import it
  at the top of the module. picomatch is small (<5ms cold load), the
  lazy-load complexity wasn't earning its keep.

Amp-Thread-ID: https://ampcode.com/threads/T-019e4c14-e44b-74fb-8597-8f0b6f205f91
